requirements.txt
pyrogram
tgcrypto
python-dotenv
youtube-dl
colorama
pydub
ffmpeg-python
aiohttp
aiofiles
motor
dnspython